+++ packages/thumbprint-react/components/Image/utils/get-image-service-src.js
@@ -1,0 +1,4 @@
+const getImageServiceSrc = ({ id, format, width }) =>
+    `https://d1vg1gqh4nkuns.cloudfront.net/i/${id}/width/${width}.${format}`;
+
+export default getImageServiceSrc;
